rename tests/integration to tests/components to avoid confusion with fast-integration (#14254)
