+++ tests/components/api-gateway/gateway-tests/manifests/global-commons.yaml
