+++ tests/components/api-gateway/gateway-tests/manifests/hydra-client.yaml
