+++ tests/components/api-gateway/gateway-tests/manifests/jwt-oauth-one-path-strategy.yaml
