+++ tests/components/api-gateway/gateway-tests/manifests/jwt-oauth-strategy.yaml
