+++ tests/components/api-gateway/gateway-tests/manifests/k8s/deployment.yaml
