+++ tests/components/api-gateway/gateway-tests/manifests/no_access_strategy.yaml
