+++ tests/components/api-gateway/gateway-tests/manifests/oauth-strategy.yaml
