+++ tests/components/api-gateway/gateway-tests/pkg/resource/batch.go
